refactor: better encapsulate tools config
